pyproject.toml
[project]
name = "BitTokens"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = "==3.13.*"
dependencies = [
    "accelerate==1.6.0",
    "anthropic==0.50.0",
    "bitsandbytes==0.45.5",
    "torch==2.11",
    "google-genai==1.5.0",
    "ipykernel==6.29.5",
    "ipympl==0.9.7",
    "matplotlib==3.10.1",
    "natsort==8.4.0",
    "openai==1.65.5",
    "overrides==7.7.0",
    "pandas==2.2.3",
    "polars==1.30.0",
    "pyarrow==21.0.0",
    "python-dotenv==1.0.1",
    "ruff==0.11.11",
    "scikit-learn==1.6.1",
    "seaborn==0.13.2",
    "tiktoken==0.9.0",
    "torchmetrics==1.6.2",
    "tqdm==4.67.1",
    "transformers==5.4.0",
    "typed-argument-parser==1.10.1",
    "wandb==0.19.8",
]

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true

[tool.ruff.lint]
extend-select = ["I"]
